feat: 게시글 수정 기능 구현

+++ src/pages/PostDetail.js
@@ -22,10 +22,23 @@ const PostDetail = () => {
     dispatch(getPostDetail(boardId));
   }, []);
 
-  const handleDelete = () => {
-    dispatch(deletePost(boardId));
-    alert("게시물이 삭제되었습니다");
-    navigate("/post");
+  const handleDelete = async () => {
+    if (
+      window.confirm(
+        "해당 게시물을 삭제하시겠습니까?\n삭제된 데이터는 복구할 수 없습니다."
+      )
+    ) {
+      dispatch(deletePost(boardId));
+      navigate("/post");
+      alert("게시물이 삭제되었습니다");
+    }
+    // if(a.payload.success){
+    //   alert("게시물이 삭제되었습니다");
+    // }
+  };
+
+  const handleEdit = () => {
+    navigate(`/update/${boardId}`, { state: post });
   };
 
   return (
@@ -46,12 +59,12 @@ const PostDetail = () => {
                   <LeftBtns>
                     <ContentCopyIcon className="button" />
                     <PrivateBtn>
-                      <EditIcon className="button" />
+                      <EditIcon onClick={handleEdit} className="button" />
                       <DeleteIcon onClick={handleDelete} className="button" />
                     </PrivateBtn>
                   </LeftBtns>
 
-                  <SubmitInput type="submit" value="저장" />
+                  <SubmitInput type="submit" value="수정" />
                 </ButtonContainer>
 
                 <div>
